+++ docs/js/react-dom.js
@@ -1,5 +1,5 @@
 /**
- * ReactDOM v0.14.3
+ * ReactDOM v0.14.6
  *
  * Copyright 2013-2015, Facebook, Inc.
  * All rights reserved.