+++ src/renderers/shared/reconciler/__tests__/ReactComponentLifeCycle-test.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2013-2015, Facebook, Inc.
+ * Copyright 2013-present, Facebook, Inc.
  * All rights reserved.
  *
  * This source code is licensed under the BSD-style license found in the
@@ -322,7 +322,7 @@ describe('ReactComponentLifeCycle', function() {
     ReactTestUtils.renderIntoDocument(<Component />);
     expect(console.error.argsForCall.length).toBe(1);
     expect(console.error.argsForCall[0][0]).toContain(
-      'Component is accessing getDOMNode or findDOMNode inside its render()'
+      'Component is accessing findDOMNode inside its render()'
     );
   });