+++ src/test/__tests__/ReactTestUtils-test.js
@@ -1,5 +1,5 @@
 /**
- * Copyright 2013-2015, Facebook, Inc.
+ * Copyright 2013-present, Facebook, Inc.
  * All rights reserved.
  *
  * This source code is licensed under the BSD-style license found in the
@@ -350,33 +350,6 @@ describe('ReactTestUtils', function() {
 
     // Should be document order, not mount order (which would be purple, orange)
     expect(log).toEqual(['orangepurple', 'orange', 'purple']);
-  });
-
-  it('does not warn for getDOMNode on ES6 classes', function() {
-    var Foo = React.createClass({
-      render: function() {
-        return <div />;
-      },
-    });
-
-    class Bar extends React.Component {
-      render() {
-        return <div />;
-      }
-    }
-
-    spyOn(console, 'error');
-
-    var foo = ReactTestUtils.renderIntoDocument(<Foo />);
-    expect(ReactTestUtils.isDOMComponent(foo)).toBe(false);
-
-    var bar = ReactTestUtils.renderIntoDocument(<Bar />);
-    expect(ReactTestUtils.isDOMComponent(bar)).toBe(false);
-
-    var div = ReactTestUtils.renderIntoDocument(<div />);
-    expect(ReactTestUtils.isDOMComponent(div)).toBe(true);
-
-    expect(console.error.calls.length).toBe(0);
   });
 
   it('should support injected wrapper components as DOM components', function() {